+++ packages/printer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect/printer
 
+## 0.33.37
+
+### Patch Changes
+
+- Updated dependencies [[`a047af9`](https://github.com/Effect-TS/effect/commit/a047af99447dfffc729e9c8ef0ca143537927e91)]:
+  - effect@3.4.5
+  - @effect/typeclass@0.24.37
+
 ## 0.33.36
 
 ### Patch Changes